requirements.txt
fastapi
uvicorn
openai
playwright
faiss-cpu
python-dotenv
